remove pvans utility
