+++ pvtoolsSrc/Makefile
@@ -4,7 +4,6 @@ include $(TOP)/configure/CONFIG
 
 USR_CPPFLAGS += -I$(TOP)/src/utils
 USR_CPPFLAGS += -I$(TOP)/src/remote
-USR_CPPFLAGS += -I$(TOP)/src/remoteClient
 
 PROD_DEFAULT += pvget
 pvget_SRCS += pvget.cpp
@@ -28,12 +27,6 @@ pvinfo_SRCS += pvutils.cpp
 
 PROD_DEFAULT += pvlist
 pvlist_SRCS += pvlist.cpp
-pvlist_SRCS += pvutils.cpp
-
-PROD_DEFAULT += pvans
-pvans_SRCS += pvans.cpp
-pvans_SRCS += pvutils.cpp
-pvans_SRCS += nameServerImpl.cpp
 
 PROD_LIBS += pvAccessCA pvAccess pvData ca Com
